confirmed connects to mongodb; relocate to index.js
